2024-01-03: v1.0.2
Add support for OpenCart 4.0.2.3 and PHP 8.1.
