+++ changelog.txt
@@ -1,6 +1,9 @@
 =====================================
 Date	  : Version: Description
 =====================================
-2022-04-11: v1.0.1: Hotfix whitespace issue.
+
+2024-01-03: v1.0.2 : Add support for OpenCart 4.0.2.3 and PHP 8.1.
+
+2022-04-11: v1.0.1 : Hotfix whitespace issue.
 
 2021-07-30: v1.0.0 : Initial release.
