+++ dpo/catalog/language/en-gb/checkout/dpo.php
@@ -1,0 +1,20 @@
+<?php
+/*
+ * Copyright (c) 2024 DPO Group
+ *
+ * Author: App Inlet (Pty) Ltd
+ *
+ * Released under the GNU General Public License
+ */
+
+// Heading
+$_['heading_title'] = 'Your order is %s!';
+
+// Text
+$_['text_basket']   = 'Shopping Cart';
+$_['text_checkout'] = 'Checkout';
+$_['text_success']  = 'Success';
+$_['text_customer'] = '<p>You can view your order history by going to the <a href="%s">my account</a> page and by clicking on <a href="%s">history</a>.</p><p>If your purchase has an associated download, you can go to the account <a href="%s">downloads</a> page to view them.</p><p>Please direct any questions you have to the <a href="%s">store owner</a>.</p><p>Thanks for shopping with us online!</p>';
+$_['text_guest']    = '<p>Please direct any questions you have to the <a href="%s">store owner</a>.</p><p>Thanks for shopping with us online!</p>';
+$_['text_title']        = 'DPO Pay';
+$_['text_dpo_checkout'] = 'Pay via DPO Pay';
