+++ dpo/catalog/language/en-gb/payment/dpo.php
@@ -1,0 +1,12 @@
+<?php
+/*
+ * Copyright (c) 2024 DPO Group
+ *
+ * Author: App Inlet (Pty) Ltd
+ *
+ * Released under the GNU General Public License
+ */
+
+// Text
+$_['text_title']        = 'DPO Pay';
+$_['text_dpo_checkout'] = 'Pay via DPO Pay';
