+++ dpo/install.json
@@ -1,0 +1,7 @@
+{
+	"name": "DPO Pay",
+	"code": "dpo",
+	"version": "1.0.2",
+	"author": "App Inlet",
+	"link": "https://appinlet.com"
+}
