--- upload/admin/language/en-gb/extension/payment/dpo.php
@@ -1,41 +1,0 @@
-<?php
-/*
- * Copyright (c) 2021 DPO Group
- *
- * Author: App Inlet (Pty) Ltd
- *
- * Released under the GNU General Public License
- */
-
-// Heading
-$_['heading_title'] = 'DPO Group';
-
-// Text
-$_['text_payment']   = 'Payment';
-$_['text_success']   = 'You have successfully configured the DPO Group payment module.';
-$_['text_edit']      = 'Edit DPO Group';
-$_['text_extension'] = 'Extensions';
-
-// Entry
-$_['entry_total']            = 'Total';
-$_['entry_order_status']     = 'Initial Order Status:';
-$_['entry_geo_zone']         = 'Geo Zone:';
-$_['entry_status']           = 'Status:';
-$_['entry_sort_order']       = 'Sort Order:';
-$_['text_dpo']               = '<a onclick="window.open(\'https://www.dpogroup.com/\');"><img src="view/image/payment/dpo.png" alt="DPO Group" title="DPO Group" style="border: 1px solid #EEEEEE;" /></a>';
-$_['entry_merchant_token']   = 'Merchant Token:';
-$_['entry_service_type']     = 'Service Type:';
-$_['entry_testmode']         = 'Test Mode:';
-$_['entry_success_status']   = 'Successful Order status:';
-$_['entry_failed_status']    = 'Failed Order Status:';
-$_['entry_cancelled_status'] = 'Cancelled Order Status:';
-
-// Tab
-$_['tab_general']      = 'General';
-$_['tab_order_status'] = 'Order status';
-
-// Help
-$_['help_total'] = 'The checkout total the order must reach before this payment method becomes active.';
-
-// Error
-$_['error_permission'] = 'Warning: You do not have permission to modify the DPO Group payment method!';
