--- upload/catalog/language/en-gb/extension/payment/dpo.php
@@ -1,12 +1,0 @@
-<?php
-/*
- * Copyright (c) 2021 DPO Group
- *
- * Author: App Inlet (Pty) Ltd
- *
- * Released under the GNU General Public License
- */
-
-// Text
-$_['text_title']        = 'DPO Group';
-$_['text_dpo_checkout'] = 'Pay via DPO Group';
